Moved around some files to more appropriate places. (#208)
* Deleted unnecessary files and moved contents to more appropriate space.

* Prevent form from clearing after an error. Move types to table file.
