+++ src/types/global.ts
@@ -1,18 +1,4 @@
 import { SelectableForTable } from "zapatos/schema";
 
-export interface ParaTaskCard {
-  task_id: string;
-  first_name: string;
-  last_name: string;
-  category: string;
-  description: string;
-  instructions: string | null;
-  target_max_attempts: number | null;
-  due_date: Date;
-  seen: boolean;
-  trial_count: number;
-  completed_trials: string | number | bigint | null;
-}
-
 export type Goal = SelectableForTable<"goal">;
 export type Subgoal = SelectableForTable<"subgoal">;
